move PartitionWriteManager into commons/utils/ + fix imports
