+++ src/main/java/commons/utils/PartitionWriteManager.java
@@ -1,7 +1,8 @@
-package broker;
+package commons.utils;
 
 import org.tinylog.Logger;
 import producer.RecordBatch;
+import server.internal.storage.Partition;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
