+++ src/main/java/server/internal/Broker.java
@@ -1,6 +1,7 @@
 package server.internal;
 
 import commons.FluxTopic;
+import commons.utils.PartitionWriteManager;
 import metadata.InMemoryTopicMetadataRepository;
 import metadata.Metadata;
 import org.tinylog.Logger;
